address test error, move components from js directory
