+++ portal/eproms_substudy_tailored_content/src/__tests__/App.test.js
@@ -1,9 +1,8 @@
 import { mount } from "@vue/test-utils";
-import App from "../js/components/App.vue";
+import App from "../components/App.vue";
 
 describe("Mounted App", () => {
   const wrapper = mount(App);
-
   test("has the expected html structure", () => {
     expect(wrapper.element).toMatchSnapshot();
   })
