+++ portal/eproms_substudy_tailored_content/src/components/DomainSection.vue
